# src/api/docs/swagger.yaml
definitions:
  models.Address:
    properties:
      address:
        type: string
      audit_tx_hash:
        type: string
      balance:
        type: number
      code_hash:
        type: string
      contract_type:
        type: string
      contract_updated_block:
        type: integer
      created_timestamp:
        type: integer
      deploy_tx_hash:
        type: string
      is_contract:
        type: boolean
      is_nft:
        type: boolean
      is_prep:
        type: boolean
      is_token:
        type: boolean
      log_count:
        type: integer
      name:
        type: string
      owner:
        type: string
      status:
        type: string
      symbol:
        type: string
      token_standard:
        type: string
      token_transfer_count:
        type: integer
      transaction_count:
        type: integer
      transaction_internal_count:
        type: integer
      type:
        type: string
    type: object
  models.AddressList:
    properties:
      address:
        type: string
      balance:
        type: number
      is_contract:
        type: boolean
      is_nft:
        type: boolean
      is_token:
        type: boolean
      token_standard:
        type: string
      transaction_count:
        type: integer
      type:
        type: string
    type: object
...
  models.ContractList:
    properties:
      address:
        type: string
      balance:
        type: number
      created_timestamp:
        type: integer
      is_token:
        type: boolean
      log_count:
        type: integer
      name:
        type: string
      status:
        type: string
      symbol:
        type: string
      token_standard:
        type: string
      transaction_count:
        type: integer
    type: object
...
  models.TokenTransfer:
    properties:
      block_number:
        type: integer
      block_timestamp:
        type: integer
      from_address:
        type: string
      log_index:
        type: integer
      nft_id:
        type: integer
      to_address:
        type: string
      token_contract_address:
        type: string
      token_contract_name:
        type: string
      token_contract_symbol:
        type: string
      transaction_fee:
        type: string
      transaction_hash:
        type: string
      transaction_index:
        type: integer
      value:
        type: string
      value_decimal:
        type: number
    type: object
...
  models.TransactionInternalList:
    properties:
      block_hash:
        type: string
      block_number:
        type: integer
      block_timestamp:
        type: integer
      data:
        type: string
      from_address:
        type: string
      hash:
        type: string
      method:
        type: string
      status:
        type: string
      to_address:
        type: string
      transaction_index:
        type: integer
      type:
        type: string
      value:
        type: string
    type: object
  models.TransactionList:
    properties:
      block_number:
        type: integer
      block_timestamp:
        type: integer
      data:
        type: string
      from_address:
        type: string
      hash:
        type: string
      method:
        type: string
      status:
        type: string
      to_address:
        type: string
      transaction_fee:
        type: string
      transaction_type:
        type: integer
      type:
        type: string
      value:
        type: string
      value_decimal:
        type: number
    type: object
info:
  contact: {}
  description: The icon tracker API
  title: Icon Go API
  version: "2.0"
paths:
  /api/v1/addresses:
    get:
      consumes:
      - '*/*'
      description: get list of addresses
      parameters:
      - description: Amount of records
        in: query
        name: limit
        type: integer
      - description: Skip to a record
        in: query
        name: skip
        type: integer
      - description: Find by address
        in: query
        name: address
        type: string
      - description: Contract addresses only
        in: query
        name: is_contract
        type: boolean
      - description: Token addresses only
        in: query
        name: is_token
        type: boolean
      - description: NFT addresses only
        in: query
        name: is_nft
        type: boolean
      - description: Token standard, either irc2, irc3, irc31
        in: query
        name: token_standard
        type: string
      - description: Field to sort by. name, balance, transaction_count, transaction_internal_count,
          token_transfer_count. Use leading `-` (ie -balance) for sort direction or
          omit for descending.
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AddressList'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get Addresses
      tags:
      - Addresses
  /api/v1/addresses/contracts:
    get:
      consumes:
      - '*/*'
      description: get list of contracts
      parameters:
      - description: contract name search
        in: query
        name: search
        type: string
      - description: tokens only
        in: query
        name: is_token
        type: boolean
      - description: NFTs only
        in: query
        name: is_nft
        type: boolean
      - description: token standard, one of irc2,irc3,irc31
        in: query
        name: token_standard
        type: string
      - description: contract status, one of active, rejected, or pending
        in: query
        name: status
        type: string
      - description: amount of records
        in: query
        name: limit
        type: integer
      - description: skip to a record
        in: query
        name: skip
        type: integer
      - description: Field to sort by. name, balance, transaction_count, transaction_internal_count,
          token_transfer_count. Use leading `-` (ie -balance) for sort direction or
          omit for descending.
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ContractList'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get contracts
      tags:
      - Addresses
  /api/v1/addresses/details/{address}:
    get:
      consumes:
      - '*/*'
      description: get details of an address
      parameters:
      - description: find by address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Address'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get Address Details
      tags:
      - Addresses
  /api/v1/addresses/token-addresses/{address}:
    get:
      consumes:
      - '*/*'
      description: get list of token contracts by address
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get Token Addresses
      tags:
      - Addresses
...
  /api/v1/transactions/address/{address}:
    get:
      consumes:
      - '*/*'
      description: get transactions by address
      parameters:
      - description: amount of records
        in: query
        name: limit
        type: integer
      - description: skip to a record
        in: query
        name: skip
        type: integer
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransactionList'
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get Transactions by address
      tags:
      - Transactions
...
  /api/v1/transactions/internal/address/{address}:
    get:
      consumes:
      - '*/*'
      description: Get internal transactions by address
      parameters:
      - description: amount of records
        in: query
        name: limit
        type: integer
      - description: skip to a record
        in: query
        name: skip
        type: integer
      - description: find by address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TransactionInternalList'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get internal transactions by address
      tags:
      - Transactions
...
  /api/v1/transactions/token-transfers/address/{address}:
    get:
      consumes:
      - '*/*'
      description: get historical token transfers by address
      parameters:
      - description: amount of records
        in: query
        name: limit
        type: integer
      - description: skip to a record
        in: query
        name: skip
        type: integer
      - description: find by address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TokenTransfer'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Get token transfer by address
      tags:
      - Transactions
...
swagger: "2.0"